+++ packages/odata-service-writer/src/types.ts
@@ -10,6 +10,11 @@ export interface NamespaceAlias {
     alias: string;
 }
 
+export enum ServiceType {
+    EDMX = 'edmx',
+    CDS = 'cds'
+}
+
 export interface OdataService {
     url?: string;
     client?: string;
@@ -17,6 +22,7 @@ export interface OdataService {
         name: string;
         instance?: string;
     };
+    type?: ServiceType;
     path?: string;
     version: OdataVersion;
     name?: string;